refactor: Extract RetryDropdown and BranchAction from message-actions
